src/api/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
i18n==0.2
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
pyjwt==2.9.0
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.1.0b3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.9.0rc1
urllib3==2.1.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
